Call select_Action as default command when calling tapis <resource>
