+++ src/packages/tapis/controllers/Systems.py
@@ -1,10 +1,8 @@
 import json
-import inspect
 
 from tapipy.errors import InvalidInputError
 
 from packages.tapis.TapisController import TapisController
-from utils.Prompt import prompt
 
 
 class Systems(TapisController):
@@ -199,64 +197,7 @@ class Systems(TapisController):
 
         return
 
-    def select_Action(self) -> None:
-        # Fetch all systems for the current user
-        systems = self.client.systems.getSystems()
-
-        # Prompt user to select from the system_ids
-        system_id = prompt.select("Select a system", [system.id for system in systems])
-
-        # Get the methods for this controller and remove the select_Action
-        methods = self.get_methods()
-        methods.remove("select_Action")
-
-        # Generate a dictionary where the key is the name of the method
-        # without the suffix "_Action" and the value is the method attr itself
-        op_map = {}
-        for op_name in methods:
-            op_map[op_name.replace("_Action", "")] = getattr(self, op_name)
-
-        # Prompt the user to select an operation to perform over the system with
-        # the selection system_id
-        action = prompt.select("Perform action", [ op for op, _ in op_map.items() ])
-        
-        filtered = filter(lambda system : system.id == system_id, systems)
-        system = None
-        # There will be only one result in filtered
-        for result in filtered:
-            system = result
-
-        # Get the arg spec for the operation being performed and
-        # remove "self" from the arguments
-        arg_spec = inspect.getfullargspec(op_map[action])
-        arg_spec.args.remove("self")
-
-        # Determine the keyword arguments. In the inspect module, the keyword
-        # arguments are the last elements of the args list. If there are any,
-        # their values will be found in the "defaults" property.
-        k_args = []
-        if arg_spec.defaults is not None:
-            k_args = arg_spec.args[-(len(arg_spec.defaults)):]
-
-        # Determine the positional arguments based on the number of keyword arguments
-        pos_args = []
-        if len(k_args) > 0:
-            pos_args = arg_spec.args[0:-(len(k_args))]
-        else:
-            pos_args = arg_spec.args
-        
-        # Prompt the use to provide values for the positional and keyword arguments
-        arg_vals = []
-        kwarg_vals = []
-        for arg in pos_args:
-            arg_vals.append(prompt.not_none(f"{arg}"))
-
-        i = 0
-        for arg in k_args:
-            kwarg_vals.append(prompt.not_none(f"{arg}", default=arg_spec.defaults[i]))
-            i = i + 1
-
-    def undelete_Action(self, system_id, test="43", test2="hello", test3=5) -> None:
+    def undelete_Action(self, system_id) -> None:
         """Undelete an applications that has been "soft" deleted."""
         try:
             self.client.systems.undeleteSystem(systemId=system_id)
